backend/requirements.txt
fastapi
uvicorn
gunicorn
google-adk
google-generativeai
pydantic
python-dotenv